feat(editor): Introduce `EditorLoader` component for lazy loading monaco editor (resolves #260).
