+++ src/components/Editor/EditorLoader.css
@@ -1,0 +1,6 @@
+.editor-loading {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    height: calc(100vh - var(--ylv-menu-bar-height) - var(--ylv-status-bar-height));
+}
