+++ src/components/MenuBar/NavigationBar.tsx
@@ -9,8 +9,8 @@ import SkipPrevious from "@mui/icons-material/SkipPrevious";
 
 import useUiStore from "../../stores/uiStore";
 import useViewStore from "../../stores/viewStore";
+import {ACTION_NAME} from "../../typings/actions";
 import {UI_ELEMENT} from "../../typings/states";
-import {ACTION_NAME} from "../../utils/actions";
 import {
     ignorePointerIfFastLoading,
     isDisabled,
