+++ src/components/MenuBar/PageNumInput.tsx
@@ -12,8 +12,8 @@ import {
 
 import useUiStore from "../../stores/uiStore";
 import useViewStore from "../../stores/viewStore";
+import {ACTION_NAME} from "../../typings/actions";
 import {UI_ELEMENT} from "../../typings/states";
-import {ACTION_NAME} from "../../utils/actions";
 import {
     ignorePointerIfFastLoading,
     isDisabled,
