+++ src/components/StatusBar/index.tsx
@@ -12,8 +12,8 @@ import AutoFixOffIcon from "@mui/icons-material/AutoFixOff";
 import useLogFileStore from "../../stores/logFileStore";
 import useUiStore from "../../stores/uiStore";
 import useViewStore from "../../stores/viewStore";
+import {ACTION_NAME} from "../../typings/actions";
 import {UI_ELEMENT} from "../../typings/states";
-import {ACTION_NAME} from "../../utils/actions";
 import {isDisabled} from "../../utils/states";
 import {copyPermalinkToClipboard} from "../../utils/url";
 import {togglePrettify} from "../../utils/url/urlHash";
