+++ src/utils/actions.ts
@@ -1,21 +1,8 @@
 import * as monaco from "monaco-editor/esm/vs/editor/editor.api.js";
 
+import {ACTION_NAME} from "../typings/actions";
 import {Nullable} from "../typings/common";
 
-
-enum ACTION_NAME {
-    SPECIFIC_PAGE = "specificPage",
-    FIRST_PAGE = "firstPage",
-    PREV_PAGE = "prevPage",
-    NEXT_PAGE = "nextPage",
-    LAST_PAGE = "lastPage",
-    PAGE_TOP = "pageTop",
-    PAGE_BOTTOM = "pageBottom",
-    RELOAD = "reload",
-    COPY_LOG_EVENT = "copyLogEvent",
-    TOGGLE_PRETTIFY = "togglePrettify",
-    TOGGLE_WORD_WRAP = "toggleWordWrap",
-}
 
 interface EditorAction {
     actionName: Nullable<ACTION_NAME>;
@@ -84,30 +71,6 @@ const EDITOR_ACTIONS: EditorAction[] = [
     },
 ];
 
-type NavigationActionsMap = {
-    [ACTION_NAME.SPECIFIC_PAGE]: {
-        pageNum: number;
-    };
-    [ACTION_NAME.FIRST_PAGE]: null;
-    [ACTION_NAME.PREV_PAGE]: null;
-    [ACTION_NAME.NEXT_PAGE]: null;
-    [ACTION_NAME.LAST_PAGE]: null;
-    [ACTION_NAME.RELOAD]: null;
-};
 
-type NavigationAction = {
-    [T in keyof NavigationActionsMap]:
-    {
-        code: T;
-        args: NavigationActionsMap[T];
-    }
-} [keyof NavigationActionsMap];
-
-export {
-    ACTION_NAME,
-    EDITOR_ACTIONS,
-};
-export type {
-    EditorAction,
-    NavigationAction,
-};
+export type {EditorAction};
+export {EDITOR_ACTIONS};
